WIP: T01 -- move await_guard to utils, add api->core RULE
