+++ design/specs/078-break-import-cycles-clean-wins/.gitignore
@@ -1,0 +1,2 @@
+trail.tsv
+trail-audit.md
