+++ design/specs/078-break-import-cycles-clean-wins/tasks/.gitignore
@@ -1,0 +1,1 @@
+.orchestrate-state.md
